pyproject.toml
[tool.poetry]
name = "exercise-shipping-service"
version = "0.1.0"
description = ""
authors = ["Ming-Tung Hong <mingtung.hong@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.2"
django-environ = "^0.10.0"
psycopg2-binary = "^2.9.6"
gunicorn = "^20.1.0"
djangorestframework = "^3.14.0"
django-extensions = "^3.2.3"
drf-spectacular = "^0.26.2"
django-filter = "^23.2"
celery = "^5.4.0"
redis = "^5.0.6"
django-celery-beat = "^2.6.0"
django-celery-results = "^2.5.1"
django-money = "^3.5.2"
json-log-formatter = "^1.0"
opentelemetry-api = "^1.25.0"
opentelemetry-sdk = "^1.25.0"
opentelemetry-exporter-otlp = "^1.25.0"
opentelemetry-instrumentation = "^0.46b0"
opentelemetry-instrumentation-wsgi = "^0.46b0"

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.2"
isort = "^5.12.0"
black = "^23.3.0"
flake8 = "^6.0.0"
pytest-django = "^4.5.2"
pytest-cov = "^4.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "tests.settings"
python_files = ["test_*.py", ]

[tool.black]
line-length = 120
skip-string-normalization = true
exclude = 'build/|dist/|\.eggs/|\.git/|\.pytest_cache/|\.tox/|\.?venv/'

[tool.isort]
profile = "black"
src_paths = [ "shipping_service" ]
known_first_party = [ "shipping_service" ]
line_length = 120
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
skip = ["shipping_service/shipping/migrations/",]
